updater updated

+++ App/Updater/Form1.cs
@@ -30,22 +30,27 @@ namespace Updater
                 foreach (string file in files)
                 {
                     var name = new FileInfo(file).Name;
-                    if (name != "Updater.exe")
+                    if (name != "System.IO.Compression.dll" && name != "System.Net.Http.dll" && name != "Updater.exe.config" && name != "System.IO.Compression.FileSystem.dll" && name != "Updater.exe")
                     {
                         File.Delete(file);
                     }
                 }
-                client.DownloadFile("https://github.com/Oyne/Khai-schedule/raw/Updater/KhaiSchedule.zip", @"KhaiScedule.zip");
+                client.DownloadFile("https://github.com/Oyne/Khai-schedule/raw/Updater/KhaiSchedule.zip", @"KhaiSchedule.zip");
                 string zipPath = @".\KhaiSchedule.zip";
                 string extarctPath = @".\";
                 ZipFile.ExtractToDirectory(zipPath, extarctPath);
                 File.Delete(@".\KhaiSchedule.zip");
-                Process.Start(@".\Updater.exe");
+                DialogResult result = MessageBox.Show("Приложение успешно обновлено", "Khai-schedule Update", MessageBoxButtons.OKCancel, MessageBoxIcon.Information);
+                if (result == DialogResult.OK)
+                {
+                    Process.Start(@".\App.exe");
+                }
                 this.Close();
             }
-            catch
+            catch (Exception ex)
             {
-                Process.Start(@".\Updater.exe");
+                MessageBox.Show("Произошла ошибка:" + ex.ToString() + "\nСкорее всего Вы установили программу в защищённую папку, из-за чего не удаётся установить новую версию.\nВ данном случае нужно скачать и установить новую версию самостоятельно.");
+                //Process.Start(@".\App.exe");
                 this.Close();
             }
